Conert rust to library; rename binary to webview
